Optimize the tuner
Instead of doing a full evaluation of the position each time a score is
required generate and store a trace of the evaluation function for each
position. The trace is then used to define an equation describing the
evaluation as a simple function of the tuning parameters. Then each time
a score is required evaluate this equation instead of calling the actual
evaluation function.