Don't replace best_move with move from transposition table
Previously, at the root of the tree, the move from the transposition table
was assumed to contain the PV move from the previous itertion and it was
set as the best move so far. However when using multiple threads this may
no longer be true since the move could have been overwritten by another
thread.