Modify how razoring is done
Previously the search depth for all moves of the position where reduced by one
when razoring is triggered. This commit changes so that the search drops
directly into quiscence search instead. It gives only a very small ELO gain
but it should hopefully simplify the addition of other pruing techniques or
reductions.