New time management while pondering
Instead of starting the clock when the opponent move is received the
clock is started directly when the engine send it's move. If the time
has already expired when the opponent move is received and it is a
ponderhit then the engine moves instantly. If it was not a ponder hit
then the clock is restarted.