Fix handling of castling in board_is_move_pseudo_legal()
The board_is_move_pseudo_legal() function should not consider if
the king moves in to check. The reason is because this case is not
checked by the move generator so the function may reject generated
moves. The fact that the king moves in to check is handled later
when making the move.