gennnuedata.py: Fix problem with storing 50-move counter
The 50-move counter was only stored with 6 bits which is not
enough. This bug is an artifact from the initial Stockfish
implementation. This is fixed by storing the counter with 7 bits
instead.

Also a Stockfish compatibility mode is introduced where the issue is
fixed in a backwards compatible manner instead. Run with the --sf flag
to enable this mode.

Additionally the padding byte is changed from 0x00 to 0xFF in order
to match the output produced by the convert command in Stockfish.